Switch to uint from int as quantities cannot be negative

+++ contracts/common/TokenOperations.sol
@@ -30,11 +30,11 @@ contract TokenOperations {
         address _token,
         address _sender,
         address _receiver,
-        int256 _amount
+        uint256 _amount
     ) internal returns (int256 responseCode) {
         bool isTransferSuccessful = isContractSendingTokens(_sender)
-            ? IERC20(_token).transfer(_receiver, uint256(_amount))
-            : IERC20(_token).transferFrom(_sender, _receiver, uint256(_amount));
+            ? IERC20(_token).transfer(_receiver, _amount)
+            : IERC20(_token).transferFrom(_sender, _receiver, _amount);
 
         return
             isTransferSuccessful
@@ -46,9 +46,9 @@ contract TokenOperations {
         address _token,
         address _sender,
         address _receiver,
-        int256 _amount
+        uint256 _amount
     ) internal {
-        IERC721(_token).transferFrom(_sender, _receiver, uint256(_amount));
+        IERC721(_token).transferFrom(_sender, _receiver, _amount);
     }
 
     function isContract(address _account) private view returns (bool) {
@@ -132,7 +132,7 @@ contract TokenOperations {
     function mintToken(
         IBaseHTS _baseHTS,
         address token,
-        int256 amount
+        uint256 amount
     ) internal returns (int256 responseCode, int256 newTotalSupply) {
         require(
             amount > 0,
@@ -155,7 +155,7 @@ contract TokenOperations {
     function burnToken(
         IBaseHTS _baseHTS,
         address token,
-        int256 amount
+        uint256 amount
     ) internal returns (int256 responseCode, int256 newTotalSupply) {
         require(
             amount > 0,
